Integrated some of QR code
Integrated some of the QR code. I've got the QR code to display after registering, the code is based off a random string generated when the user is registering and is stored along with their email. I'm having trouble getting the QR code to validate as the time it takes to read the database for the code is longer than the time it takes to execute the check. I've been looking into await and async functions but I've had no luck yet.